[Migration 0-1] Rename configs/app/features/ files to kebab-case
