+++ configs/app/features/address-3rd-party-widgets.ts
