+++ configs/app/features/ads-text.ts
