+++ configs/app/features/beacon-chain.ts
