+++ configs/app/features/easter-egg-badge.ts
