+++ configs/app/features/easter-egg-puzzle-badge.ts
