+++ configs/app/features/fault-proof-system.ts
