+++ configs/app/features/google-analytics.ts
