+++ configs/app/features/hot-contracts.ts
