+++ configs/app/features/index.ts
@@ -1,39 +1,39 @@
-export { default as advancedFilter } from './advancedFilter';
+export { default as advancedFilter } from './advanced-filter';
 export { default as account } from './account';
-export { default as addressVerification } from './addressVerification';
-export { default as addressMetadata } from './addressMetadata';
-export { default as address3rdPartyWidgets } from './address3rdPartyWidgets';
-export { default as adsBanner } from './adsBanner';
-export { default as adsText } from './adsText';
-export { default as apiDocs } from './apiDocs';
-export { default as beaconChain } from './beaconChain';
-export { default as bridgedTokens } from './bridgedTokens';
-export { default as blockchainInteraction } from './blockchainInteraction';
+export { default as addressVerification } from './address-verification';
+export { default as addressMetadata } from './address-metadata';
+export { default as address3rdPartyWidgets } from './address-3rd-party-widgets';
+export { default as adsBanner } from './ads-banner';
+export { default as adsText } from './ads-text';
+export { default as apiDocs } from './api-docs';
+export { default as beaconChain } from './beacon-chain';
+export { default as bridgedTokens } from './bridged-tokens';
+export { default as blockchainInteraction } from './blockchain-interaction';
 export { default as celo } from './celo';
-export { default as crossChainTxs } from './crossChainTxs';
-export { default as csvExport } from './csvExport';
-export { default as dataAvailability } from './dataAvailability';
-export { default as deFiDropdown } from './deFiDropdown';
-export { default as easterEggBadge } from './easterEggBadge';
-export { default as easterEggPuzzleBadge } from './easterEggPuzzleBadge';
-export { default as externalTxs } from './externalTxs';
-export { default as faultProofSystem } from './faultProofSystem';
+export { default as crossChainTxs } from './cross-chain-txs';
+export { default as csvExport } from './csv-export';
+export { default as dataAvailability } from './data-availability';
+export { default as deFiDropdown } from './defi-dropdown';
+export { default as easterEggBadge } from './easter-egg-badge';
+export { default as easterEggPuzzleBadge } from './easter-egg-puzzle-badge';
+export { default as externalTxs } from './external-txs';
+export { default as faultProofSystem } from './fault-proof-system';
 export { default as flashblocks } from './flashblocks';
-export { default as gasTracker } from './gasTracker';
-export { default as getGasButton } from './getGasButton';
-export { default as googleAnalytics } from './googleAnalytics';
-export { default as growthBook } from './growthBook';
-export { default as hotContracts } from './hotContracts';
+export { default as gasTracker } from './gas-tracker';
+export { default as getGasButton } from './get-gas-button';
+export { default as googleAnalytics } from './google-analytics';
+export { default as growthBook } from './growthbook';
+export { default as hotContracts } from './hot-contracts';
 export { default as marketplace } from './marketplace';
-export { default as megaEth } from './megaEth';
+export { default as megaEth } from './mega-eth';
 export { default as metasuites } from './metasuites';
 export { default as mixpanel } from './mixpanel';
-export { default as mudFramework } from './mudFramework';
+export { default as mudFramework } from './mud-framework';
 export { default as multichain } from './multichain';
-export { default as multichainButton } from './multichainButton';
-export { default as nameServices } from './nameServices';
+export { default as multichainButton } from './multichain-button';
+export { default as nameServices } from './name-services';
 export { default as pools } from './pools';
-export { default as publicTagsSubmission } from './publicTagsSubmission';
+export { default as publicTagsSubmission } from './public-tags-submission';
 export { default as rewards } from './rewards';
 export { default as rollbar } from './rollbar';
 export { default as rollup } from './rollup';
@@ -42,11 +42,11 @@ export { default as sol2uml } from './sol2uml';
 export { default as stats } from './stats';
 export { default as suave } from './suave';
 export { default as tac } from './tac';
-export { default as txInterpretation } from './txInterpretation';
-export { default as userOps } from './userOps';
-export { default as addressProfileAPI } from './addressProfileAPI';
+export { default as txInterpretation } from './tx-interpretation';
+export { default as userOps } from './user-ops';
+export { default as addressProfileAPI } from './address-profile-api';
 export { default as validators } from './validators';
-export { default as verifiedTokens } from './verifiedTokens';
-export { default as web3Wallet } from './web3Wallet';
-export { default as xStarScore } from './xStarScore';
+export { default as verifiedTokens } from './verified-tokens';
+export { default as web3Wallet } from './web3-wallet';
+export { default as xStarScore } from './x-star-score';
 export { default as zetachain } from './zetachain';
