+++ configs/app/features/marketplace.ts
@@ -5,7 +5,7 @@ import apis from '../apis';
 import app from '../app';
 import chain from '../chain';
 import { getEnvValue, getExternalAssetFilePath, parseEnvJson } from '../utils';
-import blockchainInteraction from './blockchainInteraction';
+import blockchainInteraction from './blockchain-interaction';
 
 const defaultTitles: MarketplaceTitles = {
   entity_name: 'Dapp',
