+++ configs/app/features/rewards.ts
@@ -3,7 +3,7 @@ import type { Feature } from './types';
 import apis from '../apis';
 import app from '../app';
 import account from './account';
-import blockchainInteraction from './blockchainInteraction';
+import blockchainInteraction from './blockchain-interaction';
 
 const title = 'Rewards service integration';
