+++ configs/app/features/x-star-score.ts
